fix: major transitions
