+++ components/Footer/index.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { motion } from "framer-motion";
 import { useTheme } from "next-themes";
 import React, { memo, useEffect, useState } from "react";
