+++ components/ThemeProvider.tsx
@@ -1,0 +1,20 @@
+"use client";
+
+import { ThemeProvider as NextThemesProvider } from "next-themes";
+import { type ThemeProviderProps } from "next-themes/dist/types";
+import { useState, useEffect } from "react";
+
+export function ThemeProvider({ children, ...props }: ThemeProviderProps) {
+  const [mounted, setMounted] = useState(false);
+
+  // Prevent hydration mismatch by only rendering after first client-side render
+  useEffect(() => {
+    setMounted(true);
+  }, []);
+
+  if (!mounted) {
+    return <>{children}</>;
+  }
+
+  return <NextThemesProvider {...props}>{children}</NextThemesProvider>;
+}
