+++ package.json
@@ -16,18 +16,20 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@tailwindcss/postcss": "^4.0.13",
     "@tailwindcss/typography": "^0.5.15",
     "autoprefixer": "^10.4.20",
     "axios": "^1.8.1",
+    "babel-plugin-react-compiler": "^19.0.0-beta-bafa41b-20250307",
     "email-validator": "^2.0.4",
     "framer-motion": "^12.4.7",
     "next": "15.2.0",
     "next-themes": "^0.4.3",
     "nodemailer": "^6.9.16",
-    "postcss": "^8.4.49",
+    "postcss": "^8.5.3",
     "react": "18.3.1",
     "react-dom": "18.3.1",
-    "tailwindcss": "^4.0.9"
+    "tailwindcss": "^4.0.13"
   },
   "devDependencies": {
     "@types/node": "22.13.8",
