+++ postcss.config.mjs
@@ -1,5 +1,6 @@
-module.exports = {
+const config = {
   plugins: {
     "@tailwindcss/postcss": {},
   },
 };
+export default config;
