+++ src/app/about/metadata.tsx
@@ -1,0 +1,6 @@
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "About | Rishabh Singh",
+  description: "About Rishabh Singh. Get more information about me!",
+};
