+++ src/app/page.tsx
@@ -1,10 +1,11 @@
-import { motion } from "framer-motion";
-import type { NextPage } from "next";
-import Head from "next/head";
-import Image from "next/image";
-import { useRouter } from "next/router";
+"use client";
 
-const Home: NextPage = () => {
+import Head from "next/head";
+import { useRouter } from "next/navigation";
+import { motion } from "framer-motion";
+import Image from "next/image";
+
+export default function Home() {
   const router = useRouter();
 
   return (
@@ -24,7 +25,13 @@ const Home: NextPage = () => {
           transition={{ delay: 0.5 }}
           className="img container flex items-center justify-center slg:w-2/4"
         >
-          <Image src={"/assets/home.png"} alt="Home Bg" className="w-full" width={1024} height={1024} />
+          <Image
+            src={"/assets/home.png"}
+            alt="Home Bg"
+            className="w-full"
+            width={1024}
+            height={1024}
+          />
         </motion.div>
         <div className="flex flex-col justify-center slg:w-2/4">
           <motion.div
@@ -82,6 +89,4 @@ const Home: NextPage = () => {
       </div>
     </div>
   );
-};
-
-export default Home;
+}
