+++ tailwind.config.js
@@ -1,18 +1,16 @@
+/** @type {import('tailwindcss').Config} */
 module.exports = {
-  darkMode: "class",
   content: [
-    "./pages/**/*.{js,ts,jsx,tsx}",
-    "./components/**/*.{js,ts,jsx,tsx}",
+    "./src/**/*.{js,ts,jsx,tsx,mdx}",
+    "./app/**/*.{js,ts,jsx,tsx,mdx}",
+    "./pages/**/*.{js,ts,jsx,tsx,mdx}",
+    "./components/**/*.{js,ts,jsx,tsx,mdx}",
   ],
+  darkMode: "class",
   theme: {
-    typography: (theme) => ({}),
     extend: {
       colors: {
         primary: "#4361EE",
-      },
-      screens: {
-        // breakpoint at 968px - slg
-        slg: "968px",
       },
     },
   },
